mos_exec now passes arguments to cmd functions
whilst the documentation for all the cmd functions said that the ptr was to the arguments, it was not - instead it was to the token for the command.  this is now fixed and points to the arguments

(NB no cmd functions were using this argument)

`cmdECHO` and `cmdPRINTF` now use this argument pointer, rather than the global `mos_strtok_ptr` token

`extractNumber` enhanced to support flags for controlling its behaviour.  also extended its functionality to optionally support `h` suffix for hex numbers (as used by the `VDU` command.  places that used `extractNumber` such as date/time variables now make use of flags to refine argument interpretation

`mos_cmdVDU` rewritten to use `extractNumber`, and has preliminary changes to support a comma-separated list of numbers to allow it to work more like