change hotkey command to use system variables
this allows for hotkeys to be defined that use macros, as well as plain strings.  exising command updated, and will now report errors correctly

`extractString` gains a “flags” byte, and it’s argument names in `mos_sysvars.h` have also been changed to reflect it’s implementation.  a flag is now supported to prevent automatic null-termination of strings in the source buffer

`createAndInsertSystemVariable` renamed to `createOrUpdateSystemVariable` as that better reflects its function

`substituteArguments` and `getArgument` support function added to `mos_sysvars.c` which are used by hotkeys. this not only supports our existing `%s` substitution, but also `%n` to substitute in a numbered argument, and `%*n` to substitute in all arguments from the nth argument onwards