add support for command aliases
`mos_exec` now checks to see if there’s an alias defined for the current command, and if so it will attempt to use that alias.  an alias can refer to another aliased command.  aliases that go around in a loop will lead to a “Too deep” error

Aliases are system variables defined with a name `Alias$newcommand`.  they support positional-substitution of arguments