commands use search paths
`runBinFile` now understands file paths

`mos_exec` will now use a path when attempting to run a non-internal command.  if the command being run includes a path (i.e. is in the form `path:cmd` or `path:dir/to/cmd`) then it leaves the path intact, otherwise will use `run:` or `moslet:` as appropriate.  this means that the search path for commands is now dynamic