add `IfThere` and `Try` commands
`IfThere` works in a similar manner to `if`, but instead of evaluating an expression it checks for the validity of a filepath.  (Filepath is transformed before being checked, so variables may be used)

`Try` will perform the given command string, trapping any errors that may occur.  it will set `Try$ReturnCode` with the return code value for the command, and if an error occurred it sets `Try$Error` to the corresponding error string.  `Try` allows for commands that can fail to be included in script files